requirements.txt
face_recognition
opencv-python
Pillow
dlib
